// Frontend/src/app/api/generated/models/LibretranslateSupportedLanguages.ts
/* tslint:disable */
/* eslint-disable */
// @ts-nocheck
/**
 * Loom API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from "../runtime";
/**
 *
 * @export
 * @interface LibretranslateSupportedLanguages
 */
export interface LibretranslateSupportedLanguages {
    /**
     *
     * @type {string}
     * @memberof LibretranslateSupportedLanguages
     */
    code: string;
    /**
     *
     * @type {string}
     * @memberof LibretranslateSupportedLanguages
     */
    name: string;
}

/**
 * Check if a given object implements the LibretranslateSupportedLanguages interface.
 */
export function instanceOfLibretranslateSupportedLanguages(
    value: object,
): boolean {
    if (!("code" in value)) return false;
    if (!("name" in value)) return false;
    return true;
}

export function LibretranslateSupportedLanguagesFromJSON(
    json: any,
): LibretranslateSupportedLanguages {
    return LibretranslateSupportedLanguagesFromJSONTyped(json, false);
}

export function LibretranslateSupportedLanguagesFromJSONTyped(
    json: any,
    ignoreDiscriminator: boolean,
): LibretranslateSupportedLanguages {
    if (json == null) {
        return json;
    }
    return {
        code: json["code"],
        name: json["name"],
    };
}

export function LibretranslateSupportedLanguagesToJSON(
    value?: LibretranslateSupportedLanguages | null,
): any {
    if (value == null) {
        return value;
    }
    return {
        code: value["code"],
        name: value["name"],
    };
}